got some performance updates out of memoizing complicated controls such as the part type selector, and other sub-components. largely still in progress, I separated the major visual components of the unwieldly Inventory page in an effort to memoize render items that don't need to be updated all the time. It's not complete and likely several things don't work yet
